requirements.txt
astroid==1.5.3
Django==1.11.4
pytz==2017.2
gunicorn==19.7.1
dj-database-url==0.4.2
psycopg2==2.7.3
flake8==3.4.1
pylint==1.7.2
pylint-django==0.7.2
selenium==3.5.0
behave-django==0.5.0
coverage==4.4.1
sphinx==1.6.3
requests==2.18.4
pyOpenSSL==17.2.0
pytz==2017.2
cfscrape==1.9.1
raven==6.1.0
whitenoise==3.3.0
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
overpy==0.4
progressbar2==3.34.3
drf-nested-routers==0.90.0